% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projr_cat.R
\name{projr_cat_pandoc_image}
\alias{projr_cat_pandoc_image}
\title{Print out Markdown image command}
\usage{
projr_cat_pandoc_image(path, caption = NULL, args = NULL)
}
\description{
Print out Markdown image command
}
