% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projr_cat.R
\name{projr_cat_plot_bg_white}
\alias{projr_cat_plot_bg_white}
\title{Print out code to add white background to a plot}
\usage{
projr_cat_plot_bg_white()
}
\description{
Print out code to add white background to a plot
}
