% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projr_cat.R
\name{projr_cat_slurm_header}
\alias{projr_cat_slurm_header}
\title{Print out top matter for a SLURM bash scripts}
\usage{
projr_cat_slurm_header()
}
\description{
Print out top matter for a SLURM bash scripts
}
