% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projr_file_ext_r_fix.R
\name{projr_file_ext_r_fix}
\alias{projr_file_ext_r_fix}
\title{Fix R and Rmd file extensions}
\usage{
projr_file_ext_r_fix(dir_r = here::here("R"), dir_rmd = here::here())
}
\arguments{
\item{dir_r, dir_rmd}{character.
Paths to directories containing R and Rmd files, respectively}
}
\description{
Fix R and Rmd file extensions

In VS Code, R and Rmd files by default
have lower-case extensions, which is non-standard and
a bit annoying.
This fixes that.
Installs the \code{here} package if not used already.
}
