% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_extra.R
\name{projr_init_extra}
\alias{projr_init_extra}
\title{Add extra bits to project set-up}
\usage{
projr_init_extra()
}
\description{
Makes .Rprofile set up renv for HPC and to use binaries,
attaches useful functions and packages and installs
packages useful for VS Code.
}
