% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pak.R
\name{projr_install_pak_devel}
\alias{projr_install_pak_devel}
\title{Install development version of pak}
\usage{
projr_install_pak_devel()
}
\description{
Occasionally \code{renv} may install a version of
\code{pak} that results in the \verb{Cannot load glue from the private library}
error. Installing from \code{r-lib} seems to work, however:
https://github.com/r-lib/pak/issues/339#issuecomment-1108238398.
This function wraps the command to do so.
}
