% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projr_renv_dep_add.R
\name{projr_renv_dep_add}
\alias{projr_renv_dep_add}
\title{Adds a package to _dependencies.R file for renv to pick up}
\usage{
projr_renv_dep_add(pkg)
}
\arguments{
\item{pkg}{character vector.
Packages to add to _dependencies.R, and install
if not already installed.
Can use "\if{html}{\out{<remote>}}/\if{html}{\out{<repo>}}" syntax for installing from
(GitHub) remotes, as long as \code{<repo>} is also
the name of the package.}
}
\description{
Adds \code{library(<pkg>)} for each
package listed. Installs them if they're not already installed.
}
