% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{projr_rprofile_hpc_renv_setup}
\alias{projr_rprofile_hpc_renv_setup}
\title{Make .Rprofile source script to make renv use scratch directory}
\usage{
projr_rprofile_hpc_renv_setup()
}
\description{
Make .Rprofile source script to make renv use scratch directory
}
