% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{projr_rprofile_renv_repos}
\alias{projr_rprofile_renv_repos}
\title{Make .Rprofile source script to make renv use right repos}
\usage{
projr_rprofile_renv_repos()
}
\description{
Copied from
\url{https://github.com/rstudio/renv/issues/1052#issuecomment-1342567839}.
}
