% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_scripts.R
\name{projr_source_script_r}
\alias{projr_source_script_r}
\title{Source all R scripts}
\usage{
projr_source_script_r()
}
\description{
Convenience function for sourcing
all \code{R} scripts in \code{"R/"} folder.
}
