% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_project_setup.R
\name{projr_tar_pipeline_activate}
\alias{projr_tar_pipeline_activate}
\title{Activate a \code{targets} pipeline}
\usage{
projr_tar_pipeline_activate(proj_nm)
}
\arguments{
\item{proj_nm}{character.
Name of \code{targets} pipeline (project).}
}
\description{
Sources all files in \verb{R/} and
runs \verb{Sys.setenv("TAR_PROJECT" = <proj_nm>)}.
}
