% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_project_setup.R
\name{projr_tar_pipeline_create}
\alias{projr_tar_pipeline_create}
\title{Create a targets pipeline}
\usage{
projr_tar_pipeline_create(proj_nm)
}
\arguments{
\item{proj_nm}{character.
Name of project.}
}
\description{
Creates a targets pipeline
corresponding to a specific targets name.
Stores the output in the temp directory,
and stores the scripts in \code{scripts/targets}.
Cats out text to add to Rmd file.
}
