% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_project_setup.R
\name{projr_tar_project_set_up}
\alias{projr_tar_project_set_up}
\title{Set up a \code{targets} pipeline}
\usage{
projr_tar_project_set_up(proj_nm)
}
\arguments{
\item{proj_nm}{character}
}
\description{
Attaches the \code{targets} package,
configures the pipeline to exist in
\code{projr_dir_get("cache", "targets", projr_nm)}
and sets the system environment variable
\code{TAR_PROJECT} to \code{proj_nm}.
}
